requirements.txt
streamlit
google-generativeai
Pillow
streamlit-option-menu
gTTS